// src/app/apiv2/model/itemResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VatResponseDTO } from './vatResponseDTO';


export interface ItemResponseDTO { 
    item_id?: number;
    name?: string;
    price_per_unit?: number;
    quantity?: number;
    quantity_unit?: string;
    article_id?: string;
    comment?: string;
    vat?: VatResponseDTO;
    preferred_list?: ItemResponseDTO.PreferredListEnum;
    preferred_list_number?: string;
    vat_type?: ItemResponseDTO.VatTypeEnum;
}
export namespace ItemResponseDTO {
    export const PreferredListEnum = {
        Rz: 'RZ',
        Ta: 'TA'
    } as const;
    export type PreferredListEnum = typeof PreferredListEnum[keyof typeof PreferredListEnum];
    export const VatTypeEnum = {
        Netto: 'netto',
        Brutto: 'brutto'
    } as const;
    export type VatTypeEnum = typeof VatTypeEnum[keyof typeof VatTypeEnum];
}